+++ resolwe_bio/tests/test_alignment.py
@@ -1,6 +1,5 @@
 # pylint: disable=missing-docstring
-from .utils import BioProcessTestCase
-
+from .utils import skipDockerFailure, BioProcessTestCase
 
 class AlignmentProcessorTestCase(BioProcessTestCase):
 
@@ -41,6 +40,7 @@ class AlignmentProcessorTestCase(BioProcessTestCase):
         aligned_reads = self.run_processor('alignment:tophat-2-0-13', inputs)
         self.assertFiles(aligned_reads, 'stats', 'tophat_reads_report.txt')
 
+    @skipDockerFailure("Fails with: STAR: command not found")
     def test_star(self):
         genome = self.prepare_genome()
         reads = self.prepare_reads()
@@ -61,6 +61,7 @@ class AlignmentProcessorTestCase(BioProcessTestCase):
         aligned_reads = self.run_processor('alignment:star', inputs)
         self.assertFiles(aligned_reads, 'gene_counts', 'gene_counts_star.tab.gz', compression='gzip')
 
+    @skipDockerFailure("File contents mismatch for bwa_bt_reads_mapped.bam")
     def test_bwa_bt(self):
         genome = self.prepare_genome()
         reads = self.prepare_reads()
@@ -79,6 +80,7 @@ class AlignmentProcessorTestCase(BioProcessTestCase):
         self.assertFiles(aligned_reads, 'bam', 'bwa_sw_reads_mapped.bam')
         self.assertFiles(aligned_reads, 'stats', 'bwa_sw_reads_report.txt')
 
+    @skipDockerFailure("File contents mismatch for bwa_mem_reads_mapped.bam")
     def test_bwa_mem(self):
         genome = self.prepare_genome()
         reads = self.prepare_reads()