+++ resolwe_bio/tests/test_expression.py
@@ -1,10 +1,13 @@
 # pylint: disable=missing-docstring
 from resolwe.flow.models import Data
 
-from .utils import BioProcessTestCase
+from .utils import skipDockerFailure, BioProcessTestCase
 
 
 class ExpressionProcessorTestCase(BioProcessTestCase):
+
+    @skipDockerFailure("Errors with: KeyError: u'mask_file' at "
+        "cuff_exp = self.run_processor('cufflinks:-2-2-1', inputs)")
     def test_cufflinks(self):
         genome = self.prepare_genome()
         reads = self.prepare_reads()
@@ -58,6 +61,7 @@ class ExpressionProcessorTestCase(BioProcessTestCase):
         cuffnorm = self.run_processor('cuffnorm:-2-2-1', inputs)
         self.assertFiles(cuffnorm, 'expset', 'expression_set.tsv.gz', compression='gzip')
 
+    @skipDockerFailure("Fails with: ImportError: No module named biox")
     def test_expression_bcm(self):
         genome = self.prepare_genome()
         reads = self.prepare_reads()
@@ -86,6 +90,7 @@ class ExpressionProcessorTestCase(BioProcessTestCase):
         etc = self.run_processor('etc:bcm-1-0-0', inputs)
         self.assertJSON(etc, etc.output['etc'], '', 'etc.json.gz')
 
+    @skipDockerFailure("Fails with: htseq-count: command not found")
     def test_expression_htseq(self):
         genome = self.prepare_genome()
         reads = self.prepare_reads()
@@ -111,6 +116,9 @@ class ExpressionProcessorTestCase(BioProcessTestCase):
         self.assertFiles(expression, 'exp', 'reads_tpm.tab.gz', compression='gzip')
         self.assertJSON(expression, expression.output['exp_json'], '', 'expression_htseq.json.gz')
 
+    @skipDockerFailure("Errors with: ERROR: basic:json value in exp_json not "
+        "ObjectId but {u'genes': {u'DPU_G0067108': 0.0, ...}} at "
+        "mergeexpression_1 = self.run_processor('mergeexpressions', inputs)")
     def test_mergeexpression(self):
         expression_1 = self.prepare_expression(f_rc='exp_1_rc.tab.gz', f_exp='exp_1_tpm.tab.gz', f_type="TPM")
         expression_2 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz', f_type="TPM")
@@ -138,6 +146,7 @@ class ExpressionProcessorTestCase(BioProcessTestCase):
         }
         self.run_processor('mergeexpressions', inputs, Data.STATUS_ERROR)
 
+    @skipDockerFailure("Fails with: ImportError: No module named biox")
     def test_etcmerge(self):
         genome = self.prepare_genome()
         reads = self.prepare_reads()
@@ -173,6 +182,8 @@ class ExpressionProcessorTestCase(BioProcessTestCase):
         etcmerge = self.run_processor('mergeetc', inputs)
         self.assertFiles(etcmerge, "expset", "merged_etc.tab.gz", compression='gzip')
 
+    @skipDockerFailure("Errors with: KeyError: u'genome' at "
+        "cuff_exp_1 = self.run_processor('cufflinks:-2-2-1', inputs)")
     def test_ncrna(self):
         inputs = {"src": "ncRNA_sample1.bam"}
         sample_1 = self.run_processor("import:upload:mapping-bam", inputs)