+++ resolwe_bio/tests/test_gencover.py
@@ -1,8 +1,10 @@
 # pylint: disable=missing-docstring
-from .utils import BioProcessTestCase
+from .utils import skipDockerFailure, BioProcessTestCase
 
 
 class GencoverProcessorTestCase(BioProcessTestCase):
+
+    @skipDockerFailure("Fails with: pypy: No such file or directory")
     def test_coverage(self):
         genome = self.prepare_genome()
         reads = self.prepare_reads()