+++ resolwe_bio/tests/test_pca.py
@@ -1,8 +1,12 @@
 # pylint: disable=missing-docstring
-from .utils import BioProcessTestCase
+from .utils import skipDockerFailure, BioProcessTestCase
 
 
 class PcaProcessorTestCase(BioProcessTestCase):
+
+    @skipDockerFailure("Errors with: ERROR: basic:json value in exp_json not "
+        "ObjectId but {u'genes': {u'DPU_G0067108': 0.0, ...}} at "
+        "pca = self.run_processor('pca:1-0-0', inputs)")
     def test_pca(self):
         expression_1 = self.prepare_expression(f_rc='exp_1_rc.tab.gz', f_exp='exp_1_tpm.tab.gz', f_type="TPM")
         expression_2 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz', f_type="TPM")