+++ resolwe_bio/tests/test_reads_filtering.py
@@ -1,9 +1,10 @@
 # pylint: disable=missing-docstring
-from .utils import BioProcessTestCase
+from .utils import skipDockerFailure, BioProcessTestCase
 
 
 class ReadsFilteringProcessorTestCase(BioProcessTestCase):
 
+    @skipDockerFailure("Fails with: prinseq-lite.pl: command not found")
     def test_prinseq_single(self):
         reads = self.prepare_reads()
         inputs = {'reads': reads.pk}
@@ -15,6 +16,7 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
         self.assertFields(filtered_reads, 'fastqc_url.url', u'fastqc/reads_fastqc/fastqc_report.html')
         self.assertFields(filtered_reads, 'fastqc_archive.file', u'reads_fastqc.zip')
 
+    @skipDockerFailure("Fails with: prinseq-lite.pl: command not found")
     def test_prinseq_paired(self):
         inputs = {
             'src1': 'rRNA_forw.fastq.gz',
@@ -32,6 +34,8 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
         self.assertFields(filtered_reads, 'fastqc_archive.file', u'rRNA_forw_fastqc.zip')
         self.assertFields(filtered_reads, 'fastqc_archive2.file', u'rRNA_rew_fastqc.zip')
 
+    @skipDockerFailure("Errors with: KeyError: u'adapters' at "
+        "filtered_reads = self.run_processor('filtering:fastq-mcf-1.1.2.537:single-end', inputs)")
     def test_fastqmcf_single(self):
         reads = self.prepare_reads()
         inputs = {'reads': reads.pk}
@@ -43,6 +47,8 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
         self.assertFields(filtered_reads, 'fastqc_url.url', u'fastqc/reads_fastqc/fastqc_report.html')
         self.assertFields(filtered_reads, 'fastqc_archive.file', u'reads_fastqc.zip')
 
+    @skipDockerFailure("Errors with: KeyError: u'adapters' at "
+        "filtered_reads = self.run_processor('filtering:fastq-mcf-1.1.2.537:paired-end', inputs)")
     def test_fastqmcf_paired(self):
         inputs = {
             'src1': 'rRNA_forw.fastq.gz',
@@ -60,6 +66,8 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
         self.assertFields(filtered_reads, 'fastqc_archive.file', u'rRNA_forw_fastqc.zip')
         self.assertFields(filtered_reads, 'fastqc_archive2.file', u'rRNA_rew_fastqc.zip')
 
+    @skipDockerFailure("Fails with: sortmerna_database.py: error: too few "
+        "arguments (slugs_path not given)")
     def test_sortmerna_single(self):
         reads = self.prepare_reads('rRNA_forw.fastq.gz')
         inputs = {
@@ -76,6 +84,8 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
         self.assertFields(filtered_reads, 'fastqc_url.url', u'fastqc/rRNA_forw_fastqc/fastqc_report.html')
         self.assertFields(filtered_reads, 'fastqc_archive.file', u'rRNA_forw_fastqc.zip')
 
+    @skipDockerFailure("Fails with: sortmerna_database.py: error: too few "
+        "arguments (slugs_path not given)")
     def test_sortmerna_paired(self):
         inputs = {
             'src1': 'rRNA_forw.fastq.gz',