+++ resolwe_bio/tests/test_support_processors.py
@@ -1,8 +1,9 @@
 # pylint: disable=missing-docstring
-from .utils import BioProcessTestCase
+from .utils import skipDockerFailure, BioProcessTestCase
 
 
 class CompatibilityProcessorTestCase(BioProcessTestCase):
+
     def test_reference_compatibility(self):
         mapping = self.prepare_bam()
         genome = self.prepare_genome('sp_test.fasta')
@@ -12,6 +13,9 @@ class CompatibilityProcessorTestCase(BioProcessTestCase):
         compatibility_test = self.run_processor('reference_compatibility', inputs)
         self.assertFiles(compatibility_test, 'report_file', 'sp_test_compatibility_report.txt')
 
+    @skipDockerFailure("Errors with: int() argument must be a string or a "
+        "number, not 'dict' at self.assertJSON(features, "
+        "features.output['feature_location'], '', 'feature_locations.json.gz')")
     def test_feature_location(self):
         inputs = {'src': 'mm10_small.gtf.gz'}
         annotation = self.run_processor('import:upload:annotation-gtf', inputs)