+++ resolwe_bio/tests/test_upload.py
@@ -3,11 +3,13 @@ import unittest
 
 from resolwe.flow.models import Data
 
-from .utils import BioProcessTestCase
+from .utils import skipDockerFailure, BioProcessTestCase
 
 
 class UploadProcessorTestCase(BioProcessTestCase):
 
+    @skipDockerFailure("Fails with: KeyError: u'proc' at "
+        "self.assertFields(upload_bam, 'proc.error', ...)")
     def test_bam_upload(self):
         inputs = {"src": "alignment_name_sorted.bam"}
         upload_bam = self.run_processor("import:upload:mapping-bam", inputs)
@@ -33,6 +35,9 @@ class UploadProcessorTestCase(BioProcessTestCase):
         self.assertFiles(upload_bam, 'bam', 'alignment_position_sorted.bam')
         self.assertFiles(upload_bam, 'bai', 'alignment_position_sorted.bam.bai')
 
+    @skipDockerFailure("Errors with: AssertionError: slug is defined before "
+        "trying to ensure uniqueness at self.run_processor("
+        "'import:upload:expression', inputs, resolwe.flow.models.Data.STATUS_ERROR)")
     def test_upload_expression(self):
         inputs = {"exp_type": "TPM"}
         self.run_processor("import:upload:expression", inputs, Data.STATUS_ERROR)
@@ -96,6 +101,7 @@ class UploadProcessorTestCase(BioProcessTestCase):
         self.assertFields(reads, "bases", "101")
         self.assertFields(reads, "fastqc_url.url", "fastqc/rRNA_forw_fastqc/fastqc_report.html")
 
+    @skipDockerFailure("File contents mismatch for old_encoding_transformed.fastq.gz")
     def test_upload_reads_old_encoding(self):
         inputs = {"src": "old_encoding.fastq.gz"}
         reads = self.run_processor("import:upload:reads-fastq", inputs)
@@ -105,6 +111,9 @@ class UploadProcessorTestCase(BioProcessTestCase):
         self.assertFields(reads, "bases", "40")
         self.assertFields(reads, "fastqc_url.url", "fastqc/old_encoding_fastqc/fastqc_report.html")
 
+    @skipDockerFailure("Errors with: int() argument must be a string or a "
+        "number, not 'dict' at self.assertJSON(diff_exp, "
+        "diff_exp.output['volcano_plot'], '', 'deseq2_volcano_plot.json.gz')")
     def test_upload_de(self):
         inputs = {'src': 'deseq2_output.tab.gz'}
         diff_exp = self.run_processor("import:upload:diffexp", inputs)