+++ resolwe_bio/tests/test_variant_calling.py
@@ -1,9 +1,10 @@
 # pylint: disable=missing-docstring
 import unittest
-from .utils import BioProcessTestCase
+from .utils import skipDockerFailure, BioProcessTestCase
 
 
 class VariantCallingTestCase(BioProcessTestCase):
+
     def test_variant_calling_samtools(self):
         genome = self.prepare_genome('variant_calling_genome.fasta.gz')
         reads = self.prepare_reads('variant_calling_reads.fastq.gz')
@@ -18,6 +19,9 @@ class VariantCallingTestCase(BioProcessTestCase):
             return line.startswith(b"##samtoolsVersion")
         self.assertFiles(samtools_variants, 'vcf', 'variant_calling_samtools.vcf', filter=filter_version)
 
+    @skipDockerFailure("Fails with: int() argument must be a string or a "
+        "number, not 'list' at "
+        "self.run_processor('vc-gatk', inputs)")
     def test_variant_calling_gatk(self):
         genome = self.prepare_genome('variant_calling_genome.fasta.gz')
         reads = self.prepare_reads('variant_calling_reads.fastq.gz')
@@ -43,6 +47,7 @@ class VariantCallingTestCase(BioProcessTestCase):
         self.run_processor('vc-gatk', inputs)
         # NOTE: output can not be tested
 
+    @skipDockerFailure("Fails with: picard: command not found")
     def test_variant_calling_gatk_joint(self):
         genome = self.prepare_genome('variant_calling_genome.fasta.gz')
         reads = self.prepare_reads('variant_calling_reads.fastq.gz')